fix(ci): split provider routing regression coverage
