+++ anyharness/crates/anyharness-lib/src/api/router.rs
@@ -11,12 +11,12 @@ use subtle::ConstantTimeEq;
 use url::form_urlencoded;
 
 use super::http::{
-    agent_auth, agent_gateway_catalog, agents, auth as http_auth, catalogs, cowork, files, git,
-    goals, health, hosting, loops, mobility, plans, processes, product_mcp, replay, repo_roots,
-    reviews, sessions, sessions_config, sessions_events, sessions_fork, sessions_interactions,
-    sessions_lifecycle, sessions_prompt, sessions_resume, subagents, terminals, workflow_runs,
-    workspaces, workspaces_lifecycle, workspaces_purge, workspaces_setup, workspaces_worktrees,
-    worktrees,
+    agent_auth::{delete_agent_auth_state, put_agent_auth_state},
+    agent_gateway_catalog, agents, auth as http_auth, catalogs, cowork, files, git, goals, health,
+    hosting, loops, mobility, plans, processes, product_mcp, replay, repo_roots, reviews, sessions,
+    sessions_config, sessions_events, sessions_fork, sessions_interactions, sessions_lifecycle,
+    sessions_prompt, sessions_resume, subagents, terminals, workflow_runs, workspaces,
+    workspaces_lifecycle, workspaces_purge, workspaces_setup, workspaces_worktrees, worktrees,
 };
 use super::sse::sessions as sse_sessions;
 use super::ws::activity as ws_activity;
@@ -68,11 +68,8 @@ pub fn build_router(state: AppState) -> Router {
             post(agents::start_agent_login_terminal),
         )
         .route("/auth/revoked-jtis", put(http_auth::push_revoked_jtis))
-        // Agent-auth state (desktop-pushed local-surface state.json)
-        .route(
-            "/agent-auth/state",
-            put(agent_auth::put_agent_auth_state).delete(agent_auth::delete_agent_auth_state),
-        )
+        .route("/agent-auth/state", put(put_agent_auth_state))
+        .route("/agent-auth/state", delete(delete_agent_auth_state))
         // Catalogs (worker-pushed agent catalog document)
         .route("/catalogs/agents", put(catalogs::apply_agent_catalog))
         .route(
