+++ server/tests/integration/test_agent_gateway_policy_enforcement.py
@@ -162,10 +162,9 @@ class TestSelectTimeEnforcement:
             sources=[],
         )
         assert cleared.status_code == 200, cleared.text
-        assert [
-            (source["sourceKind"], source["enabled"])
-            for source in cleared.json()
-        ] == [("gateway", False)]
+        assert [(source["sourceKind"], source["enabled"]) for source in cleared.json()] == [
+            ("gateway", False)
+        ]
 
     @pytest.mark.asyncio
     async def test_disallowed_harness_still_rejects_enabled_source(
